Large Changes
* Changed file structure again sorry :(
* Added bindings for luaL_register
* Fixed luaL_openlib binding
* Added luaL_openlibs binding
* Renamed helper to util and make public
* Added try_cstr, try_rstr and reg macros
* Added util::dump_stack
* Renamed LuaL_Buffer to LuaBuffer
* Added LuaReg type
* Added is_even binary module example.
* ``Type`` is no longer a useless enum and is just a namespace with constants in it.
* Removed deprecated rstring and cstring
* Reorganized interfaces
* Add ``thiserror`` as a dependency
* Status and Mode are repr(i32) now for cast safety, will probably follow suit with Type in being namespaces with constants later.
* Shortened var signatures of bindings like 50% (Still ugly!)
Largest addition here is lua regs, to make life a lot easier. Especially with the reg! macro which lets you create them easily (And no need for a null reg at the end to tell C you're done adding functions, it does that for you.)

Most of these are breaking changes but I think this will be the last time for a while, having too many single files was getting bad.